+++ src/Exploits/Exploit.ts
@@ -29,23 +29,19 @@ export enum Exploit {
   EditSaveFile = "EditSaveFile",
 }
 
-const names: Record<Exploit, string> = {
-  Bypass: "by circumventing the ram cost of document.",
-  EditSaveFile: "by editing your save file.",
-  PrototypeTampering: "by tampering with Numbers prototype.",
-  TimeCompression: "by compressing time.",
-  Unclickable: "by clicking the unclickable.",
-  UndocumentedFunctionCall: "by looking beyond the documentation.",
-  RealityAlteration: "by altering reality to suit your whims.",
-  N00dles: "by harnessing the power of the n00dles.",
-  YoureNotMeantToAccessThis: "by accessing the dev menu.",
-  TrueRecursion: "by truly recursing.",
-  INeedARainbow: "by using the power of the rainbow.",
+export const ExploitDescription: Record<Exploit, string> = {
+  [Exploit.Bypass]: "by circumventing the ram cost of document.",
+  [Exploit.EditSaveFile]: "by editing your save file.",
+  [Exploit.PrototypeTampering]: "by tampering with Numbers prototype.",
+  [Exploit.TimeCompression]: "by compressing time.",
+  [Exploit.Unclickable]: "by clicking the unclickable.",
+  [Exploit.UndocumentedFunctionCall]: "by looking beyond the documentation.",
+  [Exploit.RealityAlteration]: "by altering reality to suit your whims.",
+  [Exploit.N00dles]: "by harnessing the power of the n00dles.",
+  [Exploit.YoureNotMeantToAccessThis]: "by accessing the dev menu.",
+  [Exploit.TrueRecursion]: "by truly recursing.",
+  [Exploit.INeedARainbow]: "by using the power of the rainbow.",
 };
-
-export function ExploitName(exploit: Exploit): string {
-  return names[exploit];
-}
 
 // Needed in case player edits save file poorly
 export function sanitizeExploits(exploits: Exploit[]): Exploit[] {
